requirements.txt
# Core LLM
llama-cpp-python==0.2.90

# Memory & Database
chromadb==0.4.24
neo4j==5.15.0
redis==5.0.1
sentence-transformers==2.2.2

# Backend Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
celery==5.3.4
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Utilities
psutil==5.9.6
pywin32==306; sys_platform == 'win32'
requests==2.31.0
beautifulsoup4==4.12.2
duckduckgo-search==4.1.0
aiofiles==23.2.1

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
structlog==23.2.0

# Model Downloads
huggingface-hub==0.19.4
